requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
Django==3.1.3
future==0.18.2
html5lib==1.1
idna==2.10
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.0.1
pycparser==2.20
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
pytz==2020.4
reportlab==3.5.55
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.0
urllib3==1.25.11
webencodings==0.5.1
xhtml2pdf==0.2.5
